reduced code, moved add and commit lines into if statement

+++ src/blueprints/log_bp.py
@@ -33,8 +33,7 @@ def get_log_entries(car_id):
             <car_id> (int)
     """
     # query the database for user car's id
-    stmt = db.select(UserCar).filter_by(user_id=get_jwt_identity()).filter_by(id=car_id)
-    user_car = db.session.scalar(stmt)
+    user_car = verify_user(car_id)
     if user_car:
         # query the database and filter logs by user_car_id
         stmt = db.select(LogEntry).filter_by(user_car_id=user_car.id)
@@ -59,8 +58,7 @@ def add_log_entry(car_id):
             <car_id> (int)
     """
     # query the database for user car's id
-    stmt = db.select(UserCar).filter_by(user_id=get_jwt_identity()).filter_by(id=car_id)
-    user_car = db.session.scalar(stmt)
+    user_car = verify_user(car_id)
     if user_car:
         # load the requeest body to the log entry schema
         log_entry_info = LogEntrySchema().load(request.json)
@@ -94,8 +92,7 @@ def update_log_entry(car_id, log_id):
             <log_id> (int)
     """
     # query the database for user car's id
-    stmt = db.select(UserCar).filter_by(user_id=get_jwt_identity()).filter_by(id=car_id)
-    user_car = db.session.scalar(stmt)
+    user_car = verify_user(car_id)
     if user_car:
         # search for the log entry
         stmt = db.select(LogEntry).filter_by(id=log_id)
@@ -130,8 +127,7 @@ def delete_log_entry(car_id, log_id):
             <log_id> (int)
     """
     # query the database for user car's id
-    stmt = db.select(UserCar).filter_by(user_id=get_jwt_identity()).filter_by(id=car_id)
-    user_car = db.session.scalar(stmt)
+    user_car = verify_user(car_id)
     if user_car:
         # search for the log entry
         stmt = db.select(LogEntry).filter_by(id=log_id)
@@ -168,8 +164,7 @@ def calculate_avg_consuption(car_id):
 
     """
     # query the database for user car's id
-    stmt = db.select(UserCar).filter_by(user_id=get_jwt_identity()).filter_by(id=car_id)
-    user_car = db.session.scalar(stmt)
+    user_car = verify_user(car_id)
     if user_car:
         # filter the log entries using the user car id, order by date added
         stmt = db.select(LogEntry).filter_by(
@@ -188,13 +183,18 @@ def calculate_avg_consuption(car_id):
                 )
             )
             check_trip = db.session.scalar(stmt)
+            print(check_trip)
             if not check_trip:
+                print('yes')
             # add a new trip
                 new_trip = Trip(
                     fuel_price= trip_info['fuel_price'],
                     distance= trip_info['distance'],
                     user_car_id= car_id
                 )
+                # add and commit the trip
+                db.session.add(new_trip)
+                db.session.commit()
             # if trip exists, assign to new_trip
             new_trip = check_trip
             # last 2 entries are in list position's 0 and 1
@@ -213,9 +213,6 @@ def calculate_avg_consuption(car_id):
             trip_fuel = avg_consumption * (trip_info['distance'] / 100)
             # estimated trip cost
             trip_cost = trip_fuel * trip_info['fuel_price']
-            # add and commit the trip
-            db.session.add(new_trip)
-            db.session.commit()
             return {
                 'avg_consumption': f"{format(avg_consumption, '.2f')} L/100km",
                 'estimated_trip_fuel': f"{format(trip_fuel, '.2f')} L",
@@ -241,13 +238,17 @@ def get_all_trips(car_id):
 
             <car_id> (int)
     """
+    # verify the user
+    user = verify_user(car_id)
     # query the database
     stmt = db.select(Trip).filter_by(user_car_id=car_id)
     all_trips = db.session.scalars(stmt).all()
-    if all_trips:
-        user_trips = [trip for trip in all_trips if trip.usercar.user_id == get_jwt_identity()]
-        return TripSchema(many=True, exclude=['usercar']).dump(user_trips)
-    abort(404, "User car has no trips")
+    if user:
+        if all_trips:
+            user_trips = [trip for trip in all_trips if trip.usercar.user_id == get_jwt_identity()]
+            return TripSchema(many=True, exclude=['usercar']).dump(user_trips)
+        abort(404, "User car has no trips")
+    abort(404, "User car not found")
 
 # delete trips
 @log_bp.route('/me/<int:car_id>/trips/<int:trip_id>', methods=['DELETE'])
@@ -264,15 +265,19 @@ def delete_trips(car_id, trip_id):
 
             <trip_id> (int)
     """
+    # verify the user
+    user = verify_user(car_id)
     # query the database
     stmt = db.select(Trip).filter_by(id=trip_id)
     trip = db.session.scalar(stmt)
-    if trip and trip.usercar.user_id == get_jwt_identity() and trip.user_car_id == car_id:
-        # delete the trip and commit
-        db.session.delete(trip)
-        db.session.commit()
-        return {'deleted': 'Trip successfully delete'}
-    abort(404, "User car trip does not exist")
+    if user:
+        if trip and trip.usercar.user_id == get_jwt_identity() and trip.user_car_id == car_id:
+            # delete the trip and commit
+            db.session.delete(trip)
+            db.session.commit()
+            return {'deleted': 'Trip successfully delete'}
+        abort(404, "User car trip does not exist")
+    abort(404, "User car not found")
 
 # update trips
 @log_bp.route('/me/<int:car_id>/trips/<int:trip_id>', methods=['PUT', 'PATCH'])
@@ -289,20 +294,24 @@ def update_trips(car_id, trip_id):
 
             <trip_id> (int)
     """
+    # verify the user
+    user = verify_user(car_id)
     # query the database
     stmt = db.select(Trip).filter_by(id=trip_id)
     trip = db.session.scalar(stmt)
-    if trip and trip.usercar.user_id == get_jwt_identity() and trip.user_car_id == car_id:
-        # load the request body to the trip schema
-        trip_info = TripSchema().load(request.json)
-        # update the trip
-        trip.fuel_price = trip_info.get('fuel_price', trip.fuel_price)
-        trip.distance = trip_info.get('distance', trip.distance)
-        trip.user_car_id = car_id
-        # commit the update
-        db.session.commit()
-        return TripSchema().dump(trip)
-    abort(404, "User car trip does not exist")
+    if user:
+        if trip and trip.usercar.user_id == get_jwt_identity() and trip.user_car_id == car_id:
+            # load the request body to the trip schema
+            trip_info = TripSchema().load(request.json)
+            # update the trip
+            trip.fuel_price = trip_info.get('fuel_price', trip.fuel_price)
+            trip.distance = trip_info.get('distance', trip.distance)
+            trip.user_car_id = car_id
+            # commit the update
+            db.session.commit()
+            return TripSchema(exclude=['usercar']).dump(trip)
+        abort(404, "User car trip does not exist")
+    abort(404, "User car not found")
 
 # expenditure summary
 @log_bp.route(
@@ -357,13 +366,11 @@ def expenditure_summary(from_day, from_month, from_year, to_day, to_month, to_ye
         if logs_for_period:
             # filter out the logs that belong to the user and car
             user_logs_for_period = [log for log in logs_for_period if log.user_car_id == car_id]
-            print(user_logs_for_period)
             # calculate the total cost for the period
             total_cost = 0
             for log in user_logs_for_period:
                 total_cost += log.fuel_price * log.fuel_quantity
-            
-            
+
             return {
                     'from': f'{from_day}-{from_month}-{from_year}',
                     'to': f'{to_day}-{to_month}-{to_year}',
